# src/models/error_response.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from models.error_response_details_inner import ErrorResponseDetailsInner


class ErrorResponse(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ErrorResponse - a model defined in OpenAPI

        error_code: The error_code of this ErrorResponse.
        message: The message of this ErrorResponse.
        details: The details of this ErrorResponse [Optional].
    """

    error_code: str = Field(alias="errorCode")
    message: str = Field(alias="message")
    details: Optional[List[ErrorResponseDetailsInner]] = Field(alias="details", default=None)

ErrorResponse.update_forward_refs()