# src/models/image_generation_response.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from models.images_data import ImagesData
from models.operation_status import OperationStatus


class ImageGenerationResponse(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ImageGenerationResponse - a model defined in OpenAPI

        operation: The operation of this ImageGenerationResponse [Optional].
        data: The data of this ImageGenerationResponse [Optional].
    """

    operation: Optional[OperationStatus] = Field(alias="operation", default=None)
    data: Optional[ImagesData] = Field(alias="data", default=None)

ImageGenerationResponse.update_forward_refs()