# src/models/image_manipulation_response.py
# coding: utf-8

from __future__ import annotations
from datetime import date, datetime  # noqa: F401

import re  # noqa: F401
from typing import Any, Dict, List, Optional  # noqa: F401

from pydantic import AnyUrl, BaseModel, EmailStr, Field, validator  # noqa: F401
from models.images_data import ImagesData
from models.operation_status import OperationStatus


class ImageManipulationResponse(BaseModel):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.

    ImageManipulationResponse - a model defined in OpenAPI

        operation: The operation of this ImageManipulationResponse [Optional].
        data: The data of this ImageManipulationResponse [Optional].
    """

    operation: Optional[OperationStatus] = Field(alias="operation", default=None)
    data: Optional[ImagesData] = Field(alias="data", default=None)

ImageManipulationResponse.update_forward_refs()